+++ CheeseQuest2/game/menu/GameMenu.java
@@ -1,8 +1,8 @@
 package game.menu;
 
-import org.apache.commons.lang.WordUtils;
 import game.system.*;
 import game.object.room.*;
+import game.object.item.collectable.Collectable;
 
 /**
  * Manages all in game actions
@@ -19,7 +19,7 @@ public class GameMenu extends Menu {
 
     @Override
     public void outputPrompt() {
-        look();
+        outputLook();
     }
     public static GameMenu getInstance() {
         if (instance == null) {
@@ -47,7 +47,41 @@ public class GameMenu extends Menu {
      * Adjacent rooms
      */
     public void outputLook() {
-        outputTitle(room.getSingleName());
+        room.outputDescription();
+        // outputItems();
+
+    }
+
+    /**
+     * Output information about all items present in room
+     */
+    // public void outputItems() {
+    //     Inventory<Collectable> inv = room.getInventory();
+    //     if (!inv.isEmpty()) {
+    //         output("There is ");
+    //         for (Collectable item : inv.getItemSet()) {
+    //             int count = inv.getCount(item);
+    //             if (count == 1) {
+    //
+    //             }
+    //             if (startsWithVowel(item.getSingleName())) {
+    //
+    //             }
+    //         }
+    //     }
+    // }
+    public boolean startsWithVowel(String string) {
+        switch(Character.toLowerCase(string.charAt(0))) {
+            case 'a':
+            case 'e':
+            case 'i':
+            case 'o':
+            case 'u':
+                return true;
+                break;
+            default:
+                return false;
+        }
     }
     // The Northmen say the world will come to an end. The sun will grow black. The earth will sink into the sea. The stars will disappear. Fire and water will meet. Flames will play against the sky. The heavens and earth and all the world will be burned. All the gods will be dead, and the warriors of Valhalla, and people everywhere. Senua, prepare yourself for Ragnarok, for it is nigh.
     // The Northmen say the world will come to an end. The sun will grow black. The earth will sink into the sea. The stars will disappear. Fire and water will meet. Flames will play against the sky. The heavens and earth and all the world will be burned. All the gods will be dead, the warriors of Valhalla, and people everywhere. Prepare yourself for Ragnarok, for it is nigh.