+++ CheeseQuest2/game/object/item/collectable/consumable/food/FunnelCake.java
@@ -1,6 +1,9 @@
 package game.object.item.collectable.consumable.food;
 
 public class FunnelCake extends Food {
+
+    private static final long serialVersionUID = 1L;
+
     public FunnelCake() {
         addSingleName(new String[] {"funnel cake","cake","funnel"});
         addPluralName(new String[] {"funnel cakes","cakes","funnels"});