+++ CheeseQuest2/game/object/item/collectable/consumable/food/Pie.java
@@ -1,6 +1,11 @@
 package game.object.item.collectable.consumable.food;
 
 public class Pie extends Food {
+
+    private static final long serialVersionUID = 1L;
+
+    private static Pie instance;
+
     public Pie() {
         addSingleName(new String[] {"chicken pot pie", "chicken pie", "pie"});
         addPluralName(new String[] {"chicken pot pies", "chicken pies", "pies"});
@@ -8,5 +13,12 @@ public class Pie extends Food {
         setValue(20);
         setWeight(2);
         setHunger(50);
+    }
+
+    public static Pie getInstance() {
+        if (instance == null) {
+            instance = new Pie();
+        }
+        return instance;
     }
 }