+++ CheeseQuest2/game/object/item/collectable/equipable/chest/Chest.java
@@ -1,5 +1,7 @@
 package game.object.item.collectable.equipable.chest;
 
+import game.object.item.collectable.equipable.Equipable;
+
 public abstract class Chest extends Equipable {
 
 }