+++ CheeseQuest2/game/object/item/collectable/equipable/head/Head.java
@@ -1,4 +1,6 @@
-package game.object.item.collectable.equipable.chest;
+package game.object.item.collectable.equipable.head;
+
+import game.object.item.collectable.equipable.Equipable;
 
 public abstract class Head extends Equipable {