+++ CheeseQuest2/game/object/item/collectable/equipable/offhand/OffHand.java
@@ -1,5 +1,7 @@
 package game.object.item.collectable.equipable.offhand;
 
+import game.object.item.collectable.equipable.Equipable;
+
 public abstract class OffHand extends Equipable {
 
 }