+++ CheeseQuest2/game/object/room/TestRoom.java
@@ -6,6 +6,15 @@ public class TestRoom extends Room {
 
     public TestRoom() {
         addSingleName(new String[] {"test room"});
-        addDescription(new String[] {"It's a white box with no way out."});
+    }
+
+    @Override
+    public void outputFirstDecription() {
+        output("This is the first time you're in the ");
+        outputlnRoom("test room.");
+    }
+    @Override
+    public void outputRepeatDescription() {
+        outputln("You have already been in this room.");
     }
 }